WIP: 2c-d cleanup — 336 errors remaining (CancellationReason unified between input/view; classic InputType re-exports events::InputType; SerializedBlockListItem uses real SerializedBlock; format_credits takes f32; disable_lsp_server_for_path 2 args; FetchReviewComments uses PathBuf directly; PendingImageStub.mime_type Option<String>)

+++ app/src/settings_view/billing_and_usage/usage_history_entry.rs
@@ -1,5 +1,5 @@
 // twarp: 2c-d — AI format_credits deleted; stub.
-pub fn format_credits(_credits: i64) -> String { String::new() }
+pub fn format_credits(_credits: f32) -> String { String::new() }
 use crate::{
     settings_view::billing_and_usage_page::BillingAndUsagePageAction,
     ui_components::{blended_colors, icons::Icon},
